// frontend/src/types/rental.ts
export interface Rental {
  listing_id: string;
  listing_url?: string;
  website?: string;
  city_id: number;
  title: string;
  apartment_name: string;
  locality: string;
  property_type: string;
  bedroom: number;
  bathroom?: number;
  floor?: number;
  total_floors?: number;
  furnishing?: string;
  facing_direction?: string;
  price: number;
  deposit?: number;
  maintenance?: number;
  carpet_area?: number;
  super_builtup_area?: number;
  latitude?: number;
  longitude?: number;
  posted_by?: string;
  posted_by_name?: string;
  posted_by_contact?: string;
  description?: string;
  posted_at?: string;
  is_live?: boolean;
}